Make output state mandatory
The "output" variable is special, as we require it in all
NeuronTypes. Previously, despite it being mandatory,
it was possible for custom neuron types not to specify
the output state, which would result in confusing errors.

This also reduces the boilerplate of all neuron types having
to define the same "output" state variable.